Add a workflow to test docker run on Linux
Actions implemented as docker images are currently only
supported on Linux.

This example uses -u flag on the docker run call to avoid problems
with file permissions on the source and trap folders.

An alternative solution would be to set setgid bits on the those folders so they
inherit the group ID from the host file system. Default umask tend to
permit the group to read files by default.